+++ lib/pry.rb
@@ -159,3 +159,4 @@ require "pry/prompt"
 require "pry/inspector"
 require 'pry/object_path'
 require 'pry/output'
+require "pry/deprecate"