+++ lib/pry/commands/pry_version.rb
@@ -1,17 +1,27 @@
-class Pry
-  class Command::Version < Pry::ClassCommand
-    match 'pry-version'
-    group 'Misc'
-    description 'Show Pry version.'
+class Pry::Command::Version < Pry::ClassCommand
+  match 'pry-version'
+  group 'Misc'
+  description 'Show Pry version.'
 
-    banner <<-'BANNER'
-      Show Pry version.
-    BANNER
+  banner <<-'BANNER'
+    Show Pry version.
+  BANNER
 
-    def process
-      output.puts "Pry version: #{Pry::VERSION} on Ruby #{RUBY_VERSION}."
-    end
+  def process
+    _pry_.pager.page version_string
   end
 
-  Pry::Commands.add_command(Pry::Command::Version)
+  private
+  def version_string
+    "#{_pry_.h.bright_blue('Pry')} v#{Pry::VERSION} " \
+    "(codename: #{_pry_.h.bold(Pry::VERSION_CODENAME)}) " \
+    "running on #{_pry_.h.bright_red('Ruby')} v#{RUBY_VERSION}p#{RUBY_PATCHLEVEL}. " \
+    "#{engine_string}\n"
+  end
+
+  def engine_string
+    "Engine: #{RUBY_ENGINE}" if defined?(RUBY_ENGINE) and RUBY_ENGINE != "ruby"
+  end
+
+  Pry::Commands.add_command(self)
 end